+++ examples/todo-sqlite/src/schema/todo_schema.ts
@@ -21,7 +21,7 @@ const TodoSchema = new TodoBaseEntSchema({
     }),
     // TODO remove foreignKey here?? https://github.com/lolopinto/ent/issues/1185
     creatorID: UUIDType({
-      foreignKey: { schema: "Account", column: "ID" },
+      foreignKey: { schema: "Account", column: "id" },
     }),
     completedDate: TimestampType({ index: true, nullable: true }),
     // moving away from creatorID to assigneeID to indicate who the todo is assigned to